T813 code review changes

+++ core/src/main/java/com/rivigo/riconet/core/dto/BusinessPartnerDTO.java
@@ -2,7 +2,6 @@ package com.rivigo.riconet.core.dto;
 
 import com.rivigo.zoom.common.enums.AvailabilityStatus;
 import com.rivigo.zoom.common.enums.TacticalCreditStatus;
-import java.math.BigDecimal;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
@@ -18,7 +17,7 @@ public class BusinessPartnerDTO {
   private String legalName;
   private AvailabilityStatus availabilityStatus;
   private TacticalCreditStatus tacticalCreditStatus;
-  private BigDecimal tacticalCreditMinimumBalance;
+  private Double tacticalCreditMinimumBalance;
   private String type;
   private List<LocalityDTO> localityDTOList;
   private List<String> pincodeList;
